backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.20.0
slowapi==0.1.9
python-dotenv==1.0.1
python-multipart==0.0.9
motor==3.6.0
dnspython==2.7.0
pymongo==4.9.2